Tests with pendulum code
Courtesy of Shaun Spalding, have code for a pendulum, as I've realised the maths involved are on the same line of thought. The key here was using sine/cosine to determine acceleration.

However, I've realised the key to my issues was trying to start the boat from the "bottom of the pendulum", before it has rotated left or right. This means however I end up with no movement either way, or it acts as a pendulum on the wrong axis.

Thus, my solution to a problem created by myself is to possibly simulate what the acceleration and velocity values would be at the bottom, based on the set height of the boat.
Only issue there is that, if I want to dynamically change the height it rotates to, I'm not sure I can without somehow introducing acceleration/energy to the "pendulum".