Confirmed pendulum test works for camera angle
Did a test where I added back in the camera angle and gravity code to the pendulum object, meaning both the gravity and camera angle are equal to that of the pendulum.

Only issue was, the angle added *must* be centred around 270-degrees. For example, an angle of either 225 or 315 would have the boat rocking back and forth at 45-degrees.

Also tinkered with the gravity properties of the box, giving it some more friction to make it look a bit better, falling just as quickly, but not sliding as quickly.