Tested physics rope for picking up items
Created a physics object to track the mouse, using phys_position x and y, then created a physics rope between that and the box.

It worked rather well, but did result in the box lagging behind at a set speed, especially noticeable when the mouse was moved much quicker than this speed.

Testing physics weld next